+++ Changelog.md
@@ -8,6 +8,7 @@ Features:
  * Inline Assembly: Support some restricted tokens (return, byte, address) as identifiers in Julia mode.
  * SMT Checker: If-else branch conditions are taken into account in the SMT encoding of the program
    variables.
+ * Type Checker: Issue warning for using ``public`` visibility for interface functions.
 
 Bugfixes:
  * Parser: Disallow event declarations with no parameter list.